data_platform/requirements.txt
csv
json
logging
requests
openmeteo_requests
requests_cache
pandas
retry_requests
kafka-python
sqlalchemy
python-dotenv